+++ breakwater-core/src/test_helpers/mod.rs
@@ -1,5 +1,0 @@
-pub use dev_null_tcp_stream::*;
-pub use mock_tcp_stream::*;
-
-mod dev_null_tcp_stream;
-mod mock_tcp_stream;